Fix diff drive twist concurrency issues
Before this fix, a twist message could be received and stored one
thread, in the middle of the update() of the controller.

This would be fixed by making a copy of the shared pointer at the
beginning of the update() function, added realtime box to ensure safe
concurrent access to the pointer.